Initial attempt at a swedish dictionary from aspell.
Full disclosure, as per usual I have no knowledge of the swedish language.
However I'm curious about the characters that ended up in the list of
letters, specifically a hyphen (-). Should we remove this from each word
and also from the list of letters? Or should we remove words that contain
a hyphen, or should it be retained?