Aim for 200 words per board, and try a few times each new game to strive for this.
Because it is random how the boards are generated, it is quite difficult to
intentionally get a board with an exact word count. Hence we just randomly
generate a bunch and hope for the best.

Not very performant, because generating a board actually includes a
whole bunch of reading from disk. This can be improved.