Test char prob generator in prep for refactoring.
The goal is to make it as cheap as possible to generate lots of
candidate boards so we can choose the best each time we start
a new game. Currently generating a board mutates the probability
distributions. Copying a new probability distribution means creating
new collections and filling them.

The plan is to make the generation of boards not remove items from
the distributions, but rather just keep an index into which the
latest probability to use is. That way, copying the distribution
will mean reusing the same collections under the hood.