Make generating boards non-destructive for the probability distribution.
This allows creating multiple candidate boards from a single distribution.
In turn, it means that we don't need to read distributions from disk every
time we generate a candidate board, nor do we need to remember to take a
copy of the probability object (including a deep copy of each letters probabilities)
before using it to generate boards - which could also take some time if the
number of boards is high.